+++ src/SqlAccountRestAPI/Helpers/SqlAccountApplicationConstants.cs
@@ -1,0 +1,5 @@
+namespace SqlAccountRestAPI.Helpers;
+public class ApplicationConstants
+{
+    public const string APPLICATION_NAME = "SQLAccountRestAPI";
+}
