+++ src/SqlAccountRestAPI/Helpers/SqlAccountStockAdjustmentsHelper.cs
