+++ src/SqlAccountRestAPI/Helpers/SqlAccountStockItemTemplatesHelper.cs
